Revised version of Project-1
Revisions include updates to getRandomQuote function to use a do...while
loop, and updating of the construction of the "output" variable to be
incremental and with conditionals to check for presence of citation and
year properties in a given quote object.  Finally, this version includes
a set of functions which set timers for the quotes which (a) set an
initial timer, (b) set an auto-refresh timer, and (c) reset the
auto-refresh timer on click of button (so that you don't encounter any
quick flashes of a quote following button click due to the original
"auto" timer still running.  The css file needed to be revised because
my original code upload had modified the css based on my implementation
of dealing with missing citation and year properties. This revision
eliminates the need to change the css; so I reset it back to it's
original state.